Terminando o criar order e iniciando a criação dos pagamentos

+++ order/models.py
@@ -29,5 +29,5 @@ class Order(TimeStampedModel):
 class Item(models.Model):
     product = models.ForeignKey(Product, related_name="order_items", on_delete=models.CASCADE)
     order = models.ForeignKey(Order, related_name="items", on_delete=models.CASCADE)
-    price = models.DecimalField(max_digits=20, decimal_places=2)
+    price = models.FloatField()
     quantity = models.PositiveIntegerField(validators=[MinValueValidator(1),])
